Parallel combined index building

+++ src/multi_bible_search/memcpy_long.h
@@ -1,7 +1,8 @@
 #ifndef MEMCPY_LONG_H
 #define MEMCPY_LONG_H
-
-static inline void memcpy_long(long* dest, const long* src, size_t length) {
+#include <stdint.h>
+ 
+static inline void memcpy_long(uint32_t* dest, const uint32_t* src, size_t length) {
     while (length--) {
         *dest = *src;
         dest++;
